Cambios en la estructura del proyecto y actualización de componentes
Cambio de nombre: Renombrado el componente LandingPage a Hero.
Ajustado el estilo para asegurar que el video cubra el viewport completo.
Añadida sección "About" con nuevo contenido y diseño adaptado para ajustarse al viewport completo al hacer scroll.
