+++ src/app/dashboard/page.tsx
@@ -1,20 +1,51 @@
-"use client"
+"use client";
 
-import {Container} from "@mui/material";
-import {AppRouterCacheProvider} from "@mui/material-nextjs/v14-appRouter";
-import DynamicTables from "@/components/DynamicTables";
-import React from "react";
+import React, { useState } from 'react';
+import { useSession } from 'next-auth/react';
+import { Container, Box } from '@mui/material';
+import DynamicTables from '@/components/DynamicTables';
+import DatabaseConnectionForm from '@/components/DatabaseConnectionForm';
+import GradientBackgroundWrapper from '@/components/gradientBackground';
+import Hero from '@/components/Hero';
+import Footer from "@/components/Footer";
 
-export default function DashboardPage() {
+export default function Page() {
+    const { data: session, status } = useSession();
+    const [isConnected, setIsConnected] = useState(false);
+
+    const handleConnectionSuccess = () => {
+        setIsConnected(true);
+    };
+
+    if (status === 'loading') {
+        return <p>Loading...</p>;
+    }
 
     return (
-        <AppRouterCacheProvider>
-            <Container>
-                <h1>Grid Page </h1>
-                <p>Success Connection to the Intelligent data insight tool for your business</p>
-                <DynamicTables/>
-            </Container>
-        </AppRouterCacheProvider>);
-
-
+        <GradientBackgroundWrapper>
+            {!session ? (
+                <Box sx={{ display: 'flex', flexDirection: 'column', minHeight: '100vh' }}>
+                    <Box sx={{ flex: '1' }}>
+                        <Hero />
+                    </Box>
+                    <Footer />
+                </Box>
+            ) : (
+                <Box sx={{ display: 'flex', flexDirection: 'column', minHeight: '100vh' }}>
+                    <Container sx={{ flex: '1' }}>
+                        <h1>Grid Page</h1>
+                        {!isConnected ? (
+                            <DatabaseConnectionForm onConnectionSuccess={handleConnectionSuccess} />
+                        ) : (
+                            <>
+                                <p>Success Connection to the Intelligent data insight tool for your business</p>
+                                <DynamicTables />
+                            </>
+                        )}
+                    </Container>
+                    <Footer />
+                </Box>
+            )}
+        </GradientBackgroundWrapper>
+    );
 }
