+++ src/app/globals.css
@@ -84,7 +84,7 @@
 
 html,
 body {
-    max-width: 100vw;
+    width: 100vw;
     overflow-x: hidden;
 }
